WIP: Listing filtration logic
Rewrote all filtration components to use model binding for filtration values. This allows the root component to track the listing's filtration options in its data and ensures that filtration is all locally-scoped (unlike in the 1.0 site, where filtration was tied to the global store). Implemented all necessary UX logic for tracking dice filtration options and text search. Moved text search into a generic input component, since we'll be using that for deck listings, too. Also renamed CardListing since it was overlapping with the root-level component (and not actually descriptive of what it was, anyway).