Added cards Vuex store to cache card details for detail hovers
These will persist across the duration of the session, but always try to look up the card if it doesn't exist in the store. Given that cards virtually never change, that should allow a lot fewer requests to the server for long-running sessions (though I need to evaluate the memory requirements of this!).