Modified edit buttons to show the standard Edit button on private snapshots
I may end up removing this if it is a source of confusion, but I think people will often want to jump to editing the deck from the snapshot. It also makes it more clear the difference between editing the deck and the snapshot. Fixed a missing deck title generation point, as well, and updated the deck buttons to rely on the saved copy of the deck since the editing copy doesn't have information about being a snapshot or public.